osbuild: move `format` into `manifest`
Formats describe the serialization/deserialization of the classes in the
manifest subpackage.
