+++ osbuild/manifest/format/__init__.py
