+++ setup.py
@@ -4,7 +4,7 @@ setuptools.setup(
     name="osbuild",
     version="92",
     description="A build system for OS images",
-    packages=["osbuild", "osbuild.formats", "osbuild.util"],
+    packages=["osbuild", "osbuild.manifest.format", "osbuild.util"],
     license='Apache-2.0',
     install_requires=[
         "jsonschema",
